@@ -1,5 +1,5 @@
 from functools import partial
-from typing import Any, ClassVar, Dict, List, Literal, Optional, Tuple, Type, Union
+from typing import Any, ClassVar, Literal, Optional, Union
 
 import flax
 import flax.linen as nn
@@ -40,7 +40,7 @@ class ConstantEntropyCoef(nn.Module):
 
 
 class CrossQ(OffPolicyAlgorithmJax):
-    policy_aliases: ClassVar[Dict[str, Type[CrossQPolicy]]] = {  # type: ignore[assignment]
+    policy_aliases: ClassVar[dict[str, type[CrossQPolicy]]] = {  # type: ignore[assignment]
         "MlpPolicy": CrossQPolicy,
         "SimbaPolicy": SimbaCrossQPolicy,
         # Minimal dict support using flatten()
@@ -60,12 +60,12 @@ class CrossQ(OffPolicyAlgorithmJax):
         learning_starts: int = 100,
         batch_size: int = 256,
         gamma: float = 0.99,
-        train_freq: Union[int, Tuple[int, str]] = 1,
+        train_freq: Union[int, tuple[int, str]] = 1,
         gradient_steps: int = 1,
         policy_delay: int = 3,
         action_noise: Optional[ActionNoise] = None,
-        replay_buffer_class: Optional[Type[ReplayBuffer]] = None,
-        replay_buffer_kwargs: Optional[Dict[str, Any]] = None,
+        replay_buffer_class: Optional[type[ReplayBuffer]] = None,
+        replay_buffer_kwargs: Optional[dict[str, Any]] = None,
         ent_coef: Union[str, float] = "auto",
         target_entropy: Union[Literal["auto"], float] = "auto",
         use_sde: bool = False,
@@ -73,8 +73,8 @@ class CrossQ(OffPolicyAlgorithmJax):
         use_sde_at_warmup: bool = False,
         stats_window_size: int = 100,
         tensorboard_log: Optional[str] = None,
-        policy_kwargs: Optional[Dict[str, Any]] = None,
-        param_resets: Optional[List[int]] = None,  # List of timesteps after which to reset the params
+        policy_kwargs: Optional[dict[str, Any]] = None,
+        param_resets: Optional[list[int]] = None,  # List of timesteps after which to reset the params
         verbose: int = 0,
         seed: Optional[int] = None,
         device: str = "auto",
@@ -266,7 +266,7 @@ class CrossQ(OffPolicyAlgorithmJax):
 
         def mse_loss(
             params: flax.core.FrozenDict, batch_stats: flax.core.FrozenDict, dropout_key: flax.core.FrozenDict
-        ) -> Tuple[jax.Array, jax.Array]:
+        ) -> tuple[jax.Array, jax.Array]:
             # Joint forward pass of obs/next_obs and actions/next_state_actions to have only
             # one forward pass with shape (n_critics, 2 * batch_size, 1).
             #
@@ -328,7 +328,7 @@ class CrossQ(OffPolicyAlgorithmJax):
 
         def actor_loss(
             params: flax.core.FrozenDict, batch_stats: flax.core.FrozenDict
-        ) -> Tuple[jax.Array, Tuple[jax.Array, jax.Array]]:
+        ) -> tuple[jax.Array, tuple[jax.Array, jax.Array]]:
             dist, state_updates = actor_state.apply_fn(
                 {"params": params, "batch_stats": batch_stats},
                 observations,
@@ -424,7 +424,7 @@ class CrossQ(OffPolicyAlgorithmJax):
             },
         }
 
-        def one_update(i: int, carry: Dict[str, Any]) -> Dict[str, Any]:
+        def one_update(i: int, carry: dict[str, Any]) -> dict[str, Any]:
             # Note: this method must be defined inline because
             # `fori_loop` expect a signature fn(index, carry) -> carry
             actor_state = carry["actor_state"]
