@@ -1,4 +1,4 @@
-from typing import Any, Callable, Dict, List, Optional, Type, Union
+from typing import Any, Callable, Optional, Union
 
 import flax.linen as nn
 import jax
@@ -27,7 +27,7 @@ class TQCPolicy(BaseJaxPolicy):
         observation_space: spaces.Space,
         action_space: spaces.Box,
         lr_schedule: Schedule,
-        net_arch: Optional[Union[List[int], Dict[str, List[int]]]] = None,
+        net_arch: Optional[Union[list[int], dict[str, list[int]]]] = None,
         dropout_rate: float = 0.0,
         layer_norm: bool = False,
         top_quantiles_to_drop_per_net: int = 2,
@@ -40,14 +40,14 @@ class TQCPolicy(BaseJaxPolicy):
         use_expln: bool = False,
         clip_mean: float = 2.0,
         features_extractor_class=None,
-        features_extractor_kwargs: Optional[Dict[str, Any]] = None,
+        features_extractor_kwargs: Optional[dict[str, Any]] = None,
         normalize_images: bool = True,
         optimizer_class: Callable[..., optax.GradientTransformation] = optax.adam,
-        optimizer_kwargs: Optional[Dict[str, Any]] = None,
+        optimizer_kwargs: Optional[dict[str, Any]] = None,
         n_critics: int = 2,
         share_features_extractor: bool = False,
-        actor_class: Type[nn.Module] = SquashedGaussianActor,
-        critic_class: Type[nn.Module] = ContinuousCritic,
+        actor_class: type[nn.Module] = SquashedGaussianActor,
+        critic_class: type[nn.Module] = ContinuousCritic,
     ):
         super().__init__(
             observation_space,
@@ -183,7 +183,7 @@ class SimbaTQCPolicy(TQCPolicy):
         observation_space: spaces.Space,
         action_space: spaces.Box,
         lr_schedule: Schedule,
-        net_arch: Optional[Union[List[int], Dict[str, List[int]]]] = None,
+        net_arch: Optional[Union[list[int], dict[str, list[int]]]] = None,
         dropout_rate: float = 0,
         layer_norm: bool = False,
         top_quantiles_to_drop_per_net: int = 2,
@@ -194,14 +194,14 @@ class SimbaTQCPolicy(TQCPolicy):
         use_expln: bool = False,
         clip_mean: float = 2,
         features_extractor_class=None,
-        features_extractor_kwargs: Optional[Dict[str, Any]] = None,
+        features_extractor_kwargs: Optional[dict[str, Any]] = None,
         normalize_images: bool = True,
         optimizer_class: Callable[..., optax.GradientTransformation] = optax.adamw,
-        optimizer_kwargs: Optional[Dict[str, Any]] = None,
+        optimizer_kwargs: Optional[dict[str, Any]] = None,
         n_critics: int = 2,
         share_features_extractor: bool = False,
-        actor_class: Type[nn.Module] = SimbaSquashedGaussianActor,
-        critic_class: Type[nn.Module] = SimbaContinuousCritic,
+        actor_class: type[nn.Module] = SimbaSquashedGaussianActor,
+        critic_class: type[nn.Module] = SimbaContinuousCritic,
     ):
         super().__init__(
             observation_space,
