@@ -1,4 +1,4 @@
-from typing import Optional, Type
+from typing import Optional
 
 import flax.linen as nn
 import numpy as np
@@ -173,7 +173,7 @@ def test_dqn(tmp_path) -> None:
 
 
 @pytest.mark.parametrize("replay_buffer_class", [None, HerReplayBuffer])
-def test_dict(replay_buffer_class: Optional[Type[HerReplayBuffer]]) -> None:
+def test_dict(replay_buffer_class: Optional[type[HerReplayBuffer]]) -> None:
     env = BitFlippingEnv(n_bits=2, continuous=True)
     model = SAC("MultiInputPolicy", env, replay_buffer_class=replay_buffer_class)
 
